unfuck everything and rely on the standalone mode for testing
(bzr r697.4.34)

+++ src/PlaceEntryRemote.cpp
@@ -16,6 +16,8 @@
  * Authored by: Neil Jagdish Patel <neil.patel@canonical.com>
  */
 
+#include "config.h"
+
 #include "PlaceEntryRemote.h"
 
 #include <glib/gi18n-lib.h>
@@ -226,7 +228,6 @@ PlaceEntryRemote::SetSearch (const gchar *search, std::map<gchar*, gchar*>& hint
   GVariantBuilder *builder;
   
   builder = g_variant_builder_new (G_VARIANT_TYPE ("a{ss}"));
- // g_variant_builder_add (builder, "{ss}", "foo", "bar");
 
   /* FIXME: I'm ignoring hints because we don't use them currently */
   g_dbus_proxy_call (_proxy,
@@ -245,7 +246,7 @@ PlaceEntryRemote::SetActiveSection (guint32 section_id)
 {
   g_dbus_proxy_call (_proxy,
                      "SetActiveSection",
-                     g_variant_new ("(u)", section_id),
+                     g_variant_new ("(u)", (guint32)section_id),
                      G_DBUS_CALL_FLAGS_NONE,
                      -1,
                      NULL,
@@ -343,7 +344,7 @@ PlaceEntryRemote::Update (const gchar  *dbus_path,
 
   if (_state_changed)
     state_changed.emit ();
-
+ 
   if (_position != position)
   {
     _position = position;
@@ -369,7 +370,6 @@ PlaceEntryRemote::Update (const gchar  *dbus_path,
 
     sections_model_changed.emit ();
   }
-
 
   // FIXME: Handle place entry hints
 
@@ -399,8 +399,6 @@ PlaceEntryRemote::Update (const gchar  *dbus_path,
 
     _entry_renderer_changed = true;
   }
-  
-  
 
   // FIXME: Handle entry renderer hints
